downgrade ubuntu to 16.04
