+++ .github/workflows/main.yml
@@ -7,7 +7,7 @@ name: Build Release
 jobs:
   azure-connector:
     name: azure blob utility
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-16.04
     steps:
         - uses: actions/checkout@v2
         - name: linux i386
@@ -35,7 +35,7 @@ jobs:
 
   s3-connector:
     name: S3 connector linux amd64
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-16.04
     steps:
       - uses: actions/checkout@v2
       - name: Create tarball and upload release
@@ -47,7 +47,7 @@ jobs:
 
   s3-connector-go:
     name: s3 connector using go
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-16.04
     steps:
         - uses: actions/checkout@v2
         - name: linux i386
@@ -59,8 +59,8 @@ jobs:
             GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
             GOARCH: "386"
             GOOS: linux
-            SUBDIR: "bnr-utils/nz_s3_connector_go"
-            EXECUTABLE_NAME: "nz_s3_connector"
+            SUBDIR: "bnr-utils/nz_s3Connector"
+            EXECUTABLE_NAME: "nz_s3Connector"
         - name: linux amd64
           run: |
             cd go-release-executables
@@ -70,5 +70,5 @@ jobs:
             GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
             GOARCH: amd64
             GOOS: linux
-            SUBDIR: "bnr-utils/nz_s3_connector_go"
-            EXECUTABLE_NAME: "nz_s3_connector"
+            SUBDIR: "bnr-utils/nz_s3Connector"
+            EXECUTABLE_NAME: "nz_s3Connector"
